♻️ split record page to separate logic into custom hook
