+++ src/app/record/page.tsx
@@ -1,73 +1,21 @@
 'use client';
 
-import { useForm } from 'react-hook-form';
 import { HiPlusCircle } from 'react-icons/hi';
-import { toast } from 'sonner';
-import { DateType } from 'react-tailwindcss-datepicker';
 import dayjs from 'dayjs';
-import { zodResolver } from '@hookform/resolvers/zod';
-import { ApolloError, useMutation } from '@apollo/client';
 import { createRecordSchema } from '@/schema/record/createRecordSchema';
-import { CREATE_RECORD_MUTATION } from '@/graphql/mutation/createRecord';
 import Modal from '@/components/Dialog/Dialog';
 import RecordCard from '@/components/RecordCard/RecordCard';
 import './timepicker.css';
 import RecordForm from '@/components/RecordForm/RecordForm';
-import utc from 'dayjs/plugin/utc';
-import timezone from 'dayjs/plugin/timezone';
+import useCreateRecord from './useCreateRecord';
 
-dayjs.extend(utc);
-dayjs.extend(timezone);
-
-interface CreateRecordValues {
-  heartRate: number;
-  maxPressure: number;
-  minPressure: number;
-  observations?: string;
-  date: DateType;
-  time: string;
-}
 const defaultValues = {
   date: dayjs().format('YYYY-MM-DD'),
   time: dayjs().format('HH:mm'),
 };
 
 export default function Records() {
-  const [createRecord] = useMutation<CreateRecordValues>(CREATE_RECORD_MUTATION);
-
-  const onCreateRecord = async (values: CreateRecordValues) => {
-    const date = dayjs(values.date).toISOString().split('T')[0];
-    const [hour, minute] = values.time.split(':').map((stringTime) => Number(stringTime));
-    const datetime = dayjs(date).tz('America/Argentina/Buenos_Aires').hour(hour).minute(minute).toISOString();
-
-    const createRecordInput = {
-      date: datetime,
-      heartRate: values.heartRate,
-      maxPressure: values.maxPressure,
-      minPressure: values.minPressure,
-      observations: values?.observations,
-    };
-    try {
-      const result = await createRecord({
-        variables: {
-          createRecordInput,
-        },
-      });
-      toast.success('Record created succesfully');
-      console.log({ result });
-    } catch (error) {
-      console.log('🚀 ~ file: page.tsx:43 ~ onSignIn ~ error:', JSON.stringify(error));
-      if (error instanceof ApolloError) {
-        if (error.message === 'NetworkError when attempting to fetch resource.') {
-          toast.error('The server is not reachable');
-          return;
-        }
-      }
-
-      toast.error('An error has ocurred');
-      return;
-    }
-  };
+  const { onCreateRecord } = useCreateRecord();
 
   return (
     <main className="flex min-h-screen items-center justify-center bg-[#070B15] text-gray-100">
